// src/app/components/aspnetrolemask-view/interfaces/aspnetrolemask-view.interface.ts
export interface IaspnetrolemaskView {
      roleName: string | null;    // System.String
      roleDescription: string | null;    // System.String
      mask0: number;    // System.Int32
      mask1: number;    // System.Int32
      mask2: number;    // System.Int32
      mask3: number;    // System.Int32
      mask4: number;    // System.Int32
      mask5: number;    // System.Int32
      mask6: number;    // System.Int32
      mask7: number;    // System.Int32
      mask8: number;    // System.Int32
      mask9: number;    // System.Int32
      maskA: number;    // System.Int32
      maskB: number;    // System.Int32
      maskC: number;    // System.Int32
      maskD: number;    // System.Int32
      dask0: number;    // System.Int32
      dask1: number;    // System.Int32
      dask2: number;    // System.Int32
}